Move all wallet routes to the correct place
